+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change request metric from latency histogram to simple counter.
+
 ## [0.0.2] - 2020-12-04
 
 ### Changed